added anonymous s3 support; added online profile for tests
