+++ plugin/cog/cog-rangereader-s3/pom.xml
@@ -11,6 +11,15 @@
 
     <name>Cloud Optimized GeoTIFF S3 Range Reader</name>
 
+    <profiles>
+        <profile>
+            <id>online</id>
+            <properties>
+                <online.skip.pattern>disabled</online.skip.pattern>
+            </properties>
+        </profile>
+    </profiles>
+
     <dependencies>
         <dependency>
             <groupId>it.geosolutions.imageio-ext</groupId>
@@ -26,4 +35,22 @@
         </dependency>
     </dependencies>
 
+    <build>
+        <plugins>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-surefire-plugin</artifactId>
+                <version>2.22.1</version>
+                <configuration>
+                    <runOrder>alphabetical</runOrder>
+                    <includes>
+                        <include>**/*Test.java</include>
+                    </includes>
+                    <excludes>
+                        <exclude>${online.skip.pattern}</exclude>
+                    </excludes>
+                </configuration>
+            </plugin>
+        </plugins>
+    </build>
 </project>
