+++ plugin/cog/cog-reader/pom.xml
@@ -29,6 +29,23 @@
     <profiles>
 
         <profile>
+            <id>online</id>
+            <properties>
+                <online.skip.pattern>disabled</online.skip.pattern>
+            </properties>
+            <dependencies>
+                <dependency>
+                    <groupId>it.geosolutions.imageio-ext</groupId>
+                    <artifactId>imageio-ext-cog-rangereader-http</artifactId>
+                </dependency>
+                <dependency>
+                    <groupId>it.geosolutions.imageio-ext</groupId>
+                    <artifactId>imageio-ext-cog-rangereader-s3</artifactId>
+                </dependency>
+            </dependencies>
+        </profile>
+
+        <profile>
             <id>cog-http</id>
             <properties>
                 <jar.finalName>${project.artifactId}-http-${project.version}</jar.finalName>
@@ -75,5 +92,24 @@
         </profile>
 
     </profiles>
+
+    <build>
+        <plugins>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-surefire-plugin</artifactId>
+                <version>2.22.1</version>
+                <configuration>
+                    <runOrder>alphabetical</runOrder>
+                    <includes>
+                        <include>**/*Test.java</include>
+                    </includes>
+                    <excludes>
+                        <exclude>${online.skip.pattern}</exclude>
+                    </excludes>
+                </configuration>
+            </plugin>
+        </plugins>
+    </build>
 
 </project>
