+++ plugin/cog/cog-reader/src/test/java/it/geosolutions/imageio/tiff/CogReadOnlineTest.java
@@ -31,7 +31,7 @@ import java.io.IOException;
  * @author joshfix
  * Created on 2019-08-22
  */
-public class CogReadTest {
+public class CogReadOnlineTest {
 
     private static final String cogUrl = "https://s3-us-west-2.amazonaws.com/landsat-pds/c1/L8/153/075/LC08_L1TP_153075_20190515_20190515_01_RT/LC08_L1TP_153075_20190515_20190515_01_RT_B2.TIF";
