+++ plugin/cog/pom.xml
@@ -17,6 +17,7 @@
         <okhttp.version>4.2.0</okhttp.version>
         <aws.version>2.9.24</aws.version>
         <ehcache.version>3.4.0</ehcache.version>
+        <online.skip.pattern>**/*OnlineTest.java</online.skip.pattern>
     </properties>
 
     <modules>
